Dynamic DAT download generator - Initial stage
Created an initial version of a new DAT download generation tool heavily sourced off the generate.php code. The only difference so far is that the code does not generate files to be stored on the server, but dynamically creates the file based on whatever the DB data is.

TODOs: several, mainly add a user interface, refine DAT creation, etc.